Split scraping code into a module
